requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aom=3.6.0=h6a678d5_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py38h6a678d5_7
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py38h06a4308_0
cfitsio=3.470=h5893167_7
chardet=5.2.0=pypi_0
charls=2.2.0=h2531618_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py38h06a4308_0
cloudpickle=2.2.1=py38h06a4308_0
contourpy=1.0.5=py38hdb19cb5_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
cytoolz=0.12.2=py38h5eee18b_0
dask-core=2023.4.1=py38h06a4308_0
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
expat=2.5.0=h6a678d5_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.10.0=py38h06a4308_0
giflib=5.2.1=h5eee18b_3
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=73.1=h6a678d5_0
idna=3.6=pypi_0
imagecodecs=2023.1.23=py38hc4b7b5f_0
imageio=2.33.1=py38h06a4308_0
importlib-metadata=7.0.1=py38h06a4308_0
importlib_resources=6.1.1=py38h06a4308_1
intel-openmp=2023.1.0=hdb19cb5_46306
jpeg=9e=h5eee18b_1
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.20.1=h143b758_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libavif=0.11.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libcurl=8.5.0=h251f7ec_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
libzopfli=1.0.3=he6710b0_0
locket=1.0.0=py38h06a4308_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.7.2=py38h06a4308_0
matplotlib-base=3.7.2=py38h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
networkx=3.1=py38h06a4308_0
numpy=1.24.4=pypi_0
numpy-base=1.24.3=py38h060ed82_1
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
packaging=23.2=py38h06a4308_0
partd=1.4.1=py38h06a4308_0
pcre2=10.42=hebb0a14_0
pillow=10.2.0=py38h5eee18b_0
pip=23.3.1=py38h06a4308_0
platformdirs=3.10.0=py38h06a4308_0
ply=3.11=py38_0
pooch=1.7.0=py38h06a4308_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.10=py38h6a678d5_0
pyqt5-sip=12.13.0=py38h5eee18b_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pywavelets=1.4.1=py38h5eee18b_0
pyyaml=6.0.1=py38h5eee18b_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
requests=2.31.0=py38h06a4308_1
scikit-image=0.19.3=py38h6a678d5_1
scipy=1.10.1=py38hf6e8229_1
setuptools=68.2.2=py38h06a4308_0
sip=6.7.12=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6a678d5_1
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tifffile=2023.4.12=py38h06a4308_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=py38h06a4308_0
toolz=0.12.0=py38h06a4308_0
torch=1.11.0+cu113=pypi_0
torchaudio=0.11.0+cu113=pypi_0
torchvision=0.12.0+cu113=pypi_0
tornado=6.3.3=py38h5eee18b_0
typing-extensions=4.10.0=pypi_0
urllib3=2.2.1=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zfp=1.0.0=h6a678d5_0
zipp=3.17.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
